修复获取共享内存时共享内存大小变量使用int类型 导致获取较大共享内存时溢出，提示invalid argument的错误

+++ src/base/SysLog.c
@@ -125,7 +125,7 @@ int SysLog(int loglevel,char *format,...)
 int	getcursysloglvl(char *type)
 {
 	_sys_param	*sp = NULL;
-	int	shmsize,ret;
+	size_t	shmsize,ret;
 	shmsize	=	sizeof(_sys_param)*3;
 	int	shmid;
 	if((shmid=getshmid(3,shmsize))==-1)
